+++ Paramount.Betterclassifieds.Presentation/Scripts/_references.js
@@ -17,6 +17,6 @@
 /// <reference path="toastr-2.0.1.js" />
 /// <reference path="account/login.js" />
 /// <reference path="jquery.morphext.js" />
-/// <reference path="listings/find.js" />
 /// <reference path="jquery.cleditor.min.js" />
 /// <reference path="jquery.cleditor.js" />
+/// <reference path="listings/find.js" />